Move tests to appropriate folders, use namespaces, improve phpdoc
