+++ tests/unit/Interactions/Internal/WebDriverClickAndHoldActionTest.php
@@ -13,36 +13,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use Facebook\WebDriver\Interactions\Internal\WebDriverKeyDownAction;
+namespace Facebook\WebDriver\Interactions\Internal;
 
-class WebDriverKeyDownActionTest extends \PHPUnit_Framework_TestCase {
-  /**
-   * @type WebDriverKeyDownAction
-   */
-  private $webDriverKeyDownAction;
+use Facebook\WebDriver\Internal\WebDriverLocatable;
+use Facebook\WebDriver\WebDriverMouse;
 
-  private $webDriverKeyboard;
+class WebDriverClickAndHoldActionTest extends \PHPUnit_Framework_TestCase {
+  /** @var WebDriverClickAndHoldAction */
+  private $webDriverClickAndHoldAction;
+  /** @var WebDriverMouse|\PHPUnit_Framework_MockObject_MockObject */
   private $webDriverMouse;
+  /** @var WebDriverLocatable|\PHPUnit_Framework_MockObject_MockObject  */
   private $locationProvider;
 
   public function setUp() {
-    $this->webDriverKeyboard = $this->getMock('Facebook\WebDriver\WebDriverKeyboard');
     $this->webDriverMouse = $this->getMock('Facebook\WebDriver\WebDriverMouse');
     $this->locationProvider = $this->getMock('Facebook\WebDriver\Internal\WebDriverLocatable');
-
-    $this->webDriverKeyDownAction = new WebDriverKeyDownAction(
-      $this->webDriverKeyboard,
+    $this->webDriverClickAndHoldAction = new WebDriverClickAndHoldAction(
       $this->webDriverMouse,
       $this->locationProvider
     );
   }
 
-  public function testPerformFocusesOnElementAndSendPressKeyCommand() {
+  public function testPerformSendsMouseDownCommand() {
     $coords = $this->getMockBuilder('Facebook\WebDriver\Interactions\Internal\WebDriverCoordinates')
       ->disableOriginalConstructor()->getMock();
-    $this->webDriverMouse->expects($this->once())->method('click')->with($coords);
+    $this->webDriverMouse->expects($this->once())->method('mouseDown')->with($coords);
     $this->locationProvider->expects($this->once())->method('getCoordinates')->will($this->returnValue($coords));
-    $this->webDriverKeyboard->expects($this->once())->method('pressKey');
-    $this->webDriverKeyDownAction->perform();
+    $this->webDriverClickAndHoldAction->perform();
   }
 }
