+++ tests/unit/Interactions/Internal/WebDriverCoordinatesTest.php
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use Facebook\WebDriver\Interactions\Internal\WebDriverCoordinates;
+namespace Facebook\WebDriver\Interactions\Internal;
 
 class WebDriverCoordinatesTest extends \PHPUnit_Framework_TestCase {
   public function testConstruct() {
