+++ tests/unit/Interactions/Internal/WebDriverKeyUpActionTest.php
@@ -13,31 +13,41 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use Facebook\WebDriver\Interactions\Internal\WebDriverClickAction;
+namespace Facebook\WebDriver\Interactions\Internal;
 
-class WebDriverClickActionTest extends \PHPUnit_Framework_TestCase {
-  /**
-   * @type WebDriverClickAction
-   */
-  private $webDriverClickAction;
+use Facebook\WebDriver\Internal\WebDriverLocatable;
+use Facebook\WebDriver\WebDriverKeyboard;
+use Facebook\WebDriver\WebDriverMouse;
 
+class WebDriverKeyUpActionTest extends \PHPUnit_Framework_TestCase {
+  /** @var WebDriverKeyUpAction */
+  private $webDriverKeyUpAction;
+  /** @var WebDriverKeyboard|\PHPUnit_Framework_MockObject_MockObject */
+  private $webDriverKeyboard;
+  /** @var WebDriverMouse|\PHPUnit_Framework_MockObject_MockObject */
   private $webDriverMouse;
+  /** @var WebDriverLocatable|\PHPUnit_Framework_MockObject_MockObject  */
   private $locationProvider;
 
   public function setUp() {
+    $this->webDriverKeyboard = $this->getMock('Facebook\WebDriver\WebDriverKeyboard');
     $this->webDriverMouse = $this->getMock('Facebook\WebDriver\WebDriverMouse');
     $this->locationProvider = $this->getMock('Facebook\WebDriver\Internal\WebDriverLocatable');
-    $this->webDriverClickAction = new WebDriverClickAction(
+
+    $this->webDriverKeyUpAction = new WebDriverKeyUpAction(
+      $this->webDriverKeyboard,
       $this->webDriverMouse,
-      $this->locationProvider
+      $this->locationProvider,
+      'a'
     );
   }
 
-  public function testPerformSendsClickCommand() {
+  public function testPerformFocusesOnElementAndSendPressKeyCommand() {
     $coords = $this->getMockBuilder('Facebook\WebDriver\Interactions\Internal\WebDriverCoordinates')
       ->disableOriginalConstructor()->getMock();
     $this->webDriverMouse->expects($this->once())->method('click')->with($coords);
     $this->locationProvider->expects($this->once())->method('getCoordinates')->will($this->returnValue($coords));
-    $this->webDriverClickAction->perform();
+    $this->webDriverKeyboard->expects($this->once())->method('releaseKey')->with('a');
+    $this->webDriverKeyUpAction->perform();
   }
 }
